pyproject.toml
[project]
name = "ubs-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp==3.9.1",
    "beautifulsoup4==4.12.2",
    "configparser==6.0.0",
    "fastapi==0.109.0",
    "pandas==2.2.0",
    "playwright==1.41.0",
    "pyautogen==0.2.12",
    "pytest==7.4.3",
    "python-dotenv==1.0.0",
    "python-multipart==0.0.6",
    "reportlab==4.0.8",
    "scipy==1.12.0",
    "uvicorn==0.25.0",
]
